Add taxonomy annotations to blast tabular output

Given a tbl with a column of valid ncbi accession, the function can add taxonomic annotations for each row.

add_taxonomy_columns(
  tbl,
  ncbi_accession_colname = "ncbi_accession",
  ncbi_acc_key = NULL,
  taxonomy_level = "kingdom",
  map_superkindom = TRUE,
  batch_size = 20
)

Arguments

tbl
an object of class tbl

ncbi_accession_colname
a string (default : "ncbi_accession") denoting column name of ncbi accession.

ncbi_acc_key
user specific ENTREZ api key. Get one via taxize::use_entrez()

taxonomy_level
a string indicating level of taxonomy to be mapped. Can be one of the followings

superkingdom

kingdom

phylum

subphylum

class

subclass

infraclass

cohort

order

suborder

infraorder

superfamily

family

subfamily

genus

species

tribe

no rank

map_superkindom
logical (default TRUE). map superkingdom if kingdom not found. Valid only when taxonomy_level == "kingdom".

batch_size
The number of queries to submit at a time.

Value

a tbl.

Examples

if (FALSE) {
 f <- system.file("extdata","blast_output_01.txt" ,package = "phyloR")
 d <- readr::read_delim(f, delim ="\t" , col_names = F , comment = "#")
 colnames(d) <- phyloR::get_blast_outformat_7_colnames()
 ## add kingdom
 with_kingdom <- d %>%  dplyr::slice(1:50) %>% add_taxonomy_columns(ncbi_accession_colname ="subject_acc_ver" )
 ## add species
 with_kingdom_and_species <- with_kingdom %>% add_taxonomy_columns(ncbi_accession_colname ="subject_acc_ver",
 taxonomy_level = "species")
 dplyr::glimpse(with_kingdom_and_species)
}